VortexLattice.jl

VortexLattice.jl is a Julia package for the Vortex Lattice Method I've created for my hobbies and academic projects. It is meant to make simple the conceptual design of aircraft configurations of interest, as in the example below.

It is currently capable of obtaining:

Hinge moments;

Sectional and surface coefficients;

Full configuration coefficients;

Geometry and spatial distribution plots (using matplotlib as a back-end).

using VortexLattice

# geometry

b=1.0
croot=0.25
taper=0.5
twist=-10.0
theta_root=5.0
sweep=30.0
dihedron=5.0

# flight conditions

alpha=5.0
delta_aileron=5.0

# reference dimensions

Sref=b*croot*(1.0+taper)/2
cref=(2.0*(taper^2+taper+1.0))*croot/(3.0*(1.0+taper))
bref=b

# discretization settings

cos_dist=collect(LinRange(-pi/2, pi/2, 20))
cos_dist=[sin(e)/2+0.5 for e in cos_dist]

# airfoils (see API reference for further detail)

fplate=inviscid_flatplate()

# wing sections

sleft=Section(
    [b*tand(sweep)/2, -b/2, b*tand(dihedron)/2], 
    croot*taper; 
    incidence=theta_root+twist, 
    controls=[(:aileron, 0.2, 1.0)]
)
scenter=Section(
    [0.0, 0.0, 0.0], 
    croot; incidence=theta_root, 
    controls=[(:aileron, 0.2, 1.0), 
    (:aileron, 0.2, -1.0)]
)
sright=Section(
    [b*tand(sweep)/2, b/2, b*tand(dihedron)/2], 
    croot*taper; 
    incidence=theta_root+twist, 
    controls=[(:aileron, 0.2, -1.0)]
)

cdiscs=[cos_dist, cos_dist, cos_dist]
bdiscs=[cos_dist, cos_dist]

# surfaces

surf=Surface([sleft, scenter, sright], cdiscs, bdiscs)

# complete aircraft and solution

acft=Aircraft([surf]; Sref=Sref, bref=bref, cref=cref)

# coefficient calculations

dat=get_data(acft; alpha=alpha, control_deflections=Dict(:aileron=>delta_aileron))

# plotting (based on matplotlib wrappers)

plot_aircraft(acft)

API Reference

VortexLattice.Airfoil — Type

Struct to contain information about an airfoil

camberline_coords: matrix (shape (npts, 2)) containing airfoil camber line coordinates

CL_correct: function as $C_L=C_{L, inv}$ +CL_correct(CL)

CD_correct: function as $C_D=C_{D, inv}$+ CD_correct(CL)

VortexLattice.inviscid_flatplate — Function

Constructor for inviscid flat plate airfoil

VortexLattice.Section — Type

Struct to define wing sections

LE: position of the leading edge

chord: section chord

incidence: section incidences

afl: instance of struct Airfoil to use for viscous corrections

VortexLattice.Section — Method

Constructor for wing section. Uses a flat plate airfoil as default

LE: position of the leading edge

chord: section chord

incidence: section incidence. Defaults to zero

afl: instance of struct Airfoil to use for viscous corrections. Defaults to flat plate

controls: array of tuples indicating present control surfaces. Tuples should indicate, respectively, the control's name (Symbol), the percentage of the chord in which it's initiated and its gain in the present section

VortexLattice.Surface — Type

Struct to model a lifting surface

quadrants: vector of wing quadrants

pts: points representing panel corners (shape (nm+1, nn+1, 3))

colpts: locations of collocation points (shape (nm, nn, 3))

normals: normal vectors at each collocation point (shape (nm, nn, 3))

mtosys: local coordinate system matrix at each panel

kins: locations for vortex kins (shape (nm, nn+1, 3))

chords: chord at each panel strip

widths: width of each panel strip

quarter_chords: matrix (shape (nstrips, 3)) with quarter chord positions across the span

airfoil_ponderations: array of tuples containing two Airfoil instances and a ponderation factor. Used for viscous corrections

control_ponderations: matrix of tuples indicating, respectively, the name (Symbol) and the gain (float) for each control surface acting on each panel

mac: mean aerodynamic chord

S: surface area

VortexLattice.Surface — Method

Constructor for a lifting surface based on a list of sections (from left to right)

For arguments cdiscs and bdiscs, a usage example is:

`sects=[ (LE=[1.0, -1.0, 0.0], chord=0.5), (LE=[0.0, 0.0, 0.0], chord=1.0), (LE=[1.0, -1.0, 0.0], chord=0.5) ] bdiscs=[ [0.0, 0.5, 1.0], [0.0, 1.0] ] cdiscs=[ [0.0, 0.5, 1.0], [0.0, 0.5, 1.0], [0.0, 0.5, 1.0] ]

would result in the panel corners: [ [1.0, -1.0, 0.0] [0.5, -0.5, 0.0] [0.0, 0.0, 0.0] [1.0, 1.0, 0.0] [1.25, -1.0, 0.0] [0.875, -0.5, 0.0] [0.5, 0.0, 0.0] [1.25, 1.0, 0.0] [1.5, -1.0, 0.0] [1.25, -0.5, 0.0] [1.0, 0.0, 0.0] [1.5, 1.0, 0.0] ]`

sects: vector of sections to base the surface on

cdiscs: vector of vectors, each describing a mapping for the chordwise discretization in a wing section (ranging from zero to 1, and spaced as desired for panel edges)

bdiscs: vector of vectors, each describing a mapping for the spanwise discretization in a wing quadrant

VortexLattice.Aircraft — Type

Abstract data type to contain information about an aircraft

Sref: reference surface

cref: reference chord

bref: reference span

surfaces: array of surfaces to compose the aircraft

colpts: the totality of the mesh's control points (vector of vectors format)

normals: their respective normal vectors (vector of vectors format)

mtosys: vector of local coordinate system matrices

control_ponderations: vector of vector of tuples indicating, respectively, the name (Symbol), the hinge arm and the gain for each control surface acting on each panel

kins: their respective vortex kins (vector of tuple of vectors format)

AICM3: aerodynamic influence coefficient matrix in 3D

AICM: aerodynamic influence coefficient matrix

LU_AICM: LU-decomposed aerodynamic coefficient matrix

VortexLattice.Aircraft — Method

Constructor for Aircraft struct based on an array of surfaces

surfaces: vector of surfaces

Sref: reference surface

cref: reference chord

bref: reference span

VortexLattice.VLsolve — Function

Function to return an array of vortex intensities comprising a vortex lattice method solution for the given flight conditions

acft: the aircraft object

alpha: angle of attack

beta: sideslip angle

p: x axis (stability and control coordinate system) angular velocity

q: y axis (stability and control coordinate system) angular velocity

r: z axis (stability and control coordinate system) angular velocity

control_deflections: dictionary relating control naming (Symbols) to control deflections

return: an array with the vorticity strength in each horseshoe vortex in the mesh

return: total (influence+freestream) velocity at each collocation point

VortexLattice.get_data — Function

Function to obtain all pertinent coefficients for an aircraft

acft: Aircraft struct instance

alpha: angle of attack

beta: sideslip angle

p: x axis (stability and control coordinate system) angular velocity

q: y axis (stability and control coordinate system) angular velocity

r: z axis (stability and control coordinate system) angular velocity

control_deflections: dictionary relating control symbols and control deflections

CG: vector with coordinates for momentum calculations

VortexLattice.plot_aircraft — Function

Plot aircraft geometry using matplotlib backend

acft: the aircraft to plot

VortexLattice.plot_props — Function

Plot a property throughout an aircraft's mesh

acft: the aircraft at hand

prop: array mapping the property's value at every panel